requirements.txt
Flask
flask-restx
Jinja2
pytz
PyYAML
requests
urllib3
Werkzeug
